+++ plugins/postcss-conditional-values/src/index.ts
@@ -201,3 +201,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };