+++ plugins/postcss-focus-visible/src/index.ts
@@ -135,4 +135,5 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };