+++ plugins/postcss-logical-float-and-clear/src/index.ts
@@ -63,3 +63,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };