+++ plugins/postcss-logical-viewport-units/src/index.ts
@@ -109,3 +109,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };