+++ plugins/postcss-nesting/src/index.ts
@@ -38,3 +38,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };