+++ plugins/postcss-property-rule-prelude-list/src/index.ts
@@ -39,3 +39,4 @@ const creator: PluginCreator<pluginOptions> = () => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };