+++ plugins/postcss-scope-pseudo-class/src/index.ts
@@ -88,4 +88,5 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };