+++ plugins/postcss-syntax-descriptor-syntax-production/src/index.ts
@@ -142,3 +142,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };