+++ plugins/postcss-todo-or-die/src/index.ts
@@ -143,3 +143,4 @@ const creator: PluginCreator<never> = () => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };