+++ plugins/css-blank-pseudo/src/index.ts
@@ -147,3 +147,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };