+++ plugins/css-has-pseudo/src/index.ts
@@ -197,6 +197,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };
 
 function isWithinSupportCheck(rule: Rule): boolean {
 	let ruleParent: Container<ChildNode> | Document | undefined = rule.parent;