+++ plugins/postcss-is-pseudo-class/src/index.ts
@@ -160,4 +160,5 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };