+++ plugins/postcss-progressive-custom-properties/src/index.ts
@@ -157,3 +157,4 @@ const creator: PluginCreator<null> = () => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };