+++ plugins/postcss-selector-not/src/index.ts
@@ -74,4 +74,5 @@ const creator: PluginCreator<pluginOptions> = () => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };