+++ plugins/postcss-double-position-gradients/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Double Position Gradients
 
+### Unreleased (major)
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+
 ### 6.0.4
 
 _September 21, 2025_